pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-inference"
version = "0.1.0"
description = "Config-driven local LLM inference server using Docker"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
llm-inference = "llm_inference.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_inference"]

[tool.pytest.ini_options]
testpaths = ["tests"]
